review DB spec files organization
Trying to make the 'test architecture' a bit clearer instead of having a bit of
everything everywhere.
