+++ lib/core/test/unit/Cardano/Wallet/DB/SqliteSpec.hs
@@ -37,6 +37,8 @@ import Cardano.BM.Data.Observable
     ( ObservableInstance (..) )
 import Cardano.BM.Data.Severity
     ( Severity (..) )
+import Cardano.BM.Data.Tracer
+    ( nullTracer )
 import Cardano.BM.Setup
     ( setupTrace )
 import Cardano.BM.Trace
@@ -44,13 +46,26 @@ import Cardano.BM.Trace
 import Cardano.Crypto.Wallet
     ( XPrv )
 import Cardano.Wallet.DB
-    ( DBLayer (..), ErrWalletAlreadyExists (..), PrimaryKey (..), cleanDB )
+    ( DBLayer (..)
+    , ErrNoSuchWallet (..)
+    , ErrWalletAlreadyExists (..)
+    , PrimaryKey (..)
+    , cleanDB
+    )
+import Cardano.Wallet.DB.Arbitrary
+    ( KeyValPairs (..) )
+import Cardano.Wallet.DB.Properties
+    ( properties, withDB )
 import Cardano.Wallet.DB.Sqlite
-    ( PersistState, PersistTx, SqliteContext, newDBLayer, withDBLayer )
+    ( PersistState
+    , PersistTx
+    , SqliteContext
+    , destroyDBLayer
+    , newDBLayer
+    , withDBLayer
+    )
 import Cardano.Wallet.DB.StateMachine
     ( prop_parallel, prop_sequential )
-import Cardano.Wallet.DBSpec
-    ( dbPropertyTests, withDB )
 import Cardano.Wallet.DummyTarget.Primitive.Types
     ( DummyTarget, Tx (..), block0, genesisParameters )
 import Cardano.Wallet.Primitive.AddressDerivation
@@ -58,7 +73,7 @@ import Cardano.Wallet.Primitive.AddressDerivation
 import Cardano.Wallet.Primitive.AddressDerivation.Random
     ( RndKey (..) )
 import Cardano.Wallet.Primitive.AddressDerivation.Sequential
-    ( SeqKey (..) )
+    ( SeqKey (..), generateKeyFromSeed )
 import Cardano.Wallet.Primitive.AddressDiscovery
     ( IsOurs )
 import Cardano.Wallet.Primitive.AddressDiscovery.Random
@@ -95,15 +110,19 @@ import Control.DeepSeq
 import Control.Exception
     ( throwIO )
 import Control.Monad
-    ( replicateM_, unless )
+    ( forM_, replicateM_, unless )
+import Control.Monad.IO.Class
+    ( liftIO )
 import Control.Monad.Trans.Except
-    ( runExceptT )
+    ( ExceptT, runExceptT )
 import Crypto.Hash
     ( hash )
 import Data.ByteString
     ( ByteString )
 import Data.Coerce
     ( coerce )
+import Data.Functor
+    ( ($>) )
 import Data.Maybe
     ( mapMaybe )
 import Data.Quantity
@@ -121,13 +140,16 @@ import System.Directory
 import System.IO.Error
     ( isUserError )
 import System.IO.Temp
-    ( withSystemTempFile )
+    ( emptySystemTempFile, withSystemTempFile )
 import System.IO.Unsafe
     ( unsafePerformIO )
+import System.Random
+    ( randomRIO )
 import Test.Hspec
     ( Expectation
     , Spec
     , SpecWith
+    , before
     , beforeAll
     , beforeWith
     , describe
@@ -139,7 +161,9 @@ import Test.Hspec
     , xit
     )
 import Test.QuickCheck
-    ( Property )
+    ( Property, property, (==>) )
+import Test.QuickCheck.Monadic
+    ( monadicIO )
 
 import qualified Cardano.BM.Configuration.Model as CM
 import qualified Cardano.BM.Data.Aggregated as CM
@@ -149,32 +173,36 @@ import qualified Cardano.BM.Data.SubTrace as CM
 import qualified Cardano.Wallet.Primitive.AddressDerivation.Random as Rnd
 import qualified Cardano.Wallet.Primitive.AddressDerivation.Sequential as Seq
 import qualified Data.HashMap.Strict as HM
+import qualified Data.List as L
 import qualified Data.Map as Map
+import qualified Data.Set as Set
 import qualified Data.Text as T
 
 spec :: Spec
 spec = do
-    sqliteSpec
+    sqliteSpecSeq
     sqliteSpecRnd
     loggingSpec
-    connectionSpec
+    fileModeSpec
 
-sqliteSpec :: Spec
-sqliteSpec = withDB newMemoryDBLayer $ do
-    describe "Sqlite Simple tests (SeqState)" $
-        simpleSpec testCpSeq
-    describe "Sqlite" dbPropertyTests
+sqliteSpecSeq :: Spec
+sqliteSpecSeq = withDB newMemoryDBLayer $ do
+    describe "Sqlite Simple tests (SeqState)" $ simpleSpec testCpSeq
+    describe "Sqlite" properties
     describe "Sqlite State machine tests" $ do
         it "Sequential" (prop_sequential :: TestDBSeq -> Property)
         xit "Parallel" prop_parallel
 
 sqliteSpecRnd :: Spec
 sqliteSpecRnd = withDB newMemoryDBLayer $ do
-        describe "Sqlite simple (RndState)" $
-            simpleSpec testCpRnd
+        describe "Sqlite simple (RndState)" $ simpleSpec testCpRnd
         describe "Sqlite State machine (RndState)" $ do
             it "Sequential state machine tests"
                 (prop_sequential :: TestDBRnd -> Property)
+
+{-------------------------------------------------------------------------------
+                                 Simple Specs
+-------------------------------------------------------------------------------}
 
 simpleSpec
     :: forall s k.
@@ -184,27 +212,27 @@ simpleSpec
        , GenerateTestKey k )
     => Wallet s DummyTarget
     -> SpecWith (DBLayer IO s DummyTarget k)
-simpleSpec testCp = do
+simpleSpec cp = do
     describe "Wallet table" $ do
         it "create and list works" $ \db -> do
-            unsafeRunExceptT $ createWallet db testPk testCp testMetadata
+            unsafeRunExceptT $ createWallet db testPk cp testMetadata
             listWallets db `shouldReturn` [testPk]
 
         it "create and get meta works" $ \db -> do
             now <- getCurrentTime
             let md = testMetadata
                     { passphraseInfo = Just $ WalletPassphraseInfo now }
-            unsafeRunExceptT $ createWallet db testPk testCp md
+            unsafeRunExceptT $ createWallet db testPk cp md
             readWalletMeta db testPk `shouldReturn` Just md
 
         it "create twice is handled" $ \db -> do
-            let create' = createWallet db testPk testCp testMetadata
+            let create' = createWallet db testPk cp testMetadata
             runExceptT create' `shouldReturn` (Right ())
             runExceptT create' `shouldReturn`
                 (Left (ErrWalletAlreadyExists testWid))
 
         it "create and get private key" $ \db -> do
-            unsafeRunExceptT $ createWallet db testPk testCp testMetadata
+            unsafeRunExceptT $ createWallet db testPk cp testMetadata
             readPrivateKey db testPk `shouldReturn` Nothing
             (k, h) <- generateTestKey
             unsafeRunExceptT (putPrivateKey db testPk (k, h))
@@ -212,15 +240,15 @@ simpleSpec testCp = do
 
         let sortOrder = Descending
         it "put and read tx history" $ \db -> do
-            unsafeRunExceptT $ createWallet db testPk testCp testMetadata
+            unsafeRunExceptT $ createWallet db testPk cp testMetadata
             runExceptT (putTxHistory db testPk (Map.fromList testTxs))
                 `shouldReturn` Right ()
             readTxHistory db testPk
                 sortOrder wholeRange `shouldReturn` testTxs
 
         it "put and read tx history - regression case" $ \db -> do
-            unsafeRunExceptT $ createWallet db testPk testCp testMetadata
-            unsafeRunExceptT $ createWallet db testPk1 testCp testMetadata
+            unsafeRunExceptT $ createWallet db testPk cp testMetadata
+            unsafeRunExceptT $ createWallet db testPk1 cp testMetadata
             runExceptT (putTxHistory db testPk1 (Map.fromList testTxs))
                 `shouldReturn` Right ()
             runExceptT (removeWallet db testPk) `shouldReturn` Right ()
@@ -228,9 +256,13 @@ simpleSpec testCp = do
                 sortOrder wholeRange `shouldReturn` testTxs
 
         it "put and read checkpoint" $ \db -> do
-            unsafeRunExceptT $ createWallet db testPk testCp testMetadata
-            runExceptT (putCheckpoint db testPk testCp) `shouldReturn` Right ()
-            readCheckpoint db testPk `shouldReturn` Just testCp
+            unsafeRunExceptT $ createWallet db testPk cp testMetadata
+            runExceptT (putCheckpoint db testPk cp) `shouldReturn` Right ()
+            readCheckpoint db testPk `shouldReturn` Just cp
+
+{-------------------------------------------------------------------------------
+                                Logging Spec
+-------------------------------------------------------------------------------}
 
 loggingSpec :: Spec
 loggingSpec = withLoggingDB @(SeqState DummyTarget) @DummyTarget @SeqKey $ do
@@ -256,38 +288,6 @@ loggingSpec = withLoggingDB @(SeqState DummyTarget) @DummyTarget @SeqKey $ do
             -- T.unpack msgs `shouldContain` "runQuery monitor works"
             msgs <- findObserveDiffs <$> getLogs
             length msgs `shouldBe` count
-
-connectionSpec :: Spec
-connectionSpec = describe "Sqlite database file" $ do
-    let writeSomething db = do
-            unsafeRunExceptT $ createWallet db testPk testCpSeq testMetadata
-            listWallets db `shouldReturn` [testPk]
-        tempFilesAbsent fp = do
-            doesFileExist fp `shouldReturn` True
-            doesFileExist (fp <> "-wal") `shouldReturn` False
-            doesFileExist (fp <> "-shm") `shouldReturn` False
-        bomb = throwIO (userError "bomb")
-    it "is properly closed after withDBLayer" $
-        withTestDBFile writeSomething tempFilesAbsent
-    it "is properly closed after an exception in withDBLayer" $
-        withTestDBFile (\db -> writeSomething db >> bomb) tempFilesAbsent
-            `shouldThrow` isUserError
-
--- | Run a test action inside withDBLayer, then check assertions.
-withTestDBFile
-    :: (DBLayer IO (SeqState DummyTarget) DummyTarget SeqKey -> IO ())
-    -> (FilePath -> IO a)
-    -> IO a
-withTestDBFile action expectations = do
-    logConfig <- defaultConfigTesting
-    trace <- setupTrace (Right logConfig) "connectionSpec"
-    withSystemTempFile "spec.db" $ \fp _handle -> do
-        removeFile fp
-        withDBLayer logConfig trace (Just fp) action
-        expectations fp
-
-type TestDBSeq = DBLayer IO (SeqState DummyTarget) DummyTarget SeqKey
-type TestDBRnd = DBLayer IO (RndState DummyTarget) DummyTarget RndKey
 
 -- | Set up a DBLayer for testing, with the command context, and the logging
 -- variable.
@@ -368,8 +368,236 @@ shouldHaveLog msgs (sev, str) = unless (any match msgs) $
     match (sev', msg) = sev' == sev && str `T.isInfixOf` msg
 
 {-------------------------------------------------------------------------------
+                                File Mode Spec
+-------------------------------------------------------------------------------}
+
+type TestDBSeq = DBLayer IO (SeqState DummyTarget) DummyTarget SeqKey
+type TestDBRnd = DBLayer IO (RndState DummyTarget) DummyTarget RndKey
+
+fileModeSpec :: Spec
+fileModeSpec =  do
+    describe "Check db opening/closing" $ do
+        it "Opening and closing of db works" $ do
+            replicateM_ 25 $ do
+                db <- Just <$> temporaryDBFile
+                (ctx, _) <- newDBLayer' @(SeqState DummyTarget) @DummyTarget db
+                destroyDBLayer ctx
+
+    describe "Sqlite database file" $ do
+        let writeSomething db = do
+                unsafeRunExceptT $ createWallet db testPk testCpSeq testMetadata
+                listWallets db `shouldReturn` [testPk]
+            tempFilesAbsent fp = do
+                doesFileExist fp `shouldReturn` True
+                doesFileExist (fp <> "-wal") `shouldReturn` False
+                doesFileExist (fp <> "-shm") `shouldReturn` False
+            bomb = throwIO (userError "bomb")
+        it "is properly closed after withDBLayer" $
+            withTestDBFile writeSomething tempFilesAbsent
+        it "is properly closed after an exception in withDBLayer" $
+            withTestDBFile (\db -> writeSomething db >> bomb) tempFilesAbsent
+                `shouldThrow` isUserError
+
+    before temporaryDBFile $
+        describe "Check db reading/writing from/to file and cleaning" $ do
+
+        it "create and list wallet works" $ \f -> do
+            (ctx, db) <- newDBLayer' (Just f)
+            unsafeRunExceptT $ createWallet db testPk testCp testMetadata
+            destroyDBLayer ctx
+            testOpeningCleaning f listWallets [testPk] []
+
+        it "create and get meta works" $ \f -> do
+            (ctx, db) <- newDBLayer' (Just f)
+            now <- getCurrentTime
+            let meta = testMetadata
+                   { passphraseInfo = Just $ WalletPassphraseInfo now }
+            unsafeRunExceptT $ createWallet db testPk testCp meta
+            destroyDBLayer ctx
+            testOpeningCleaning f (`readWalletMeta` testPk) (Just meta) Nothing
+
+        it "create and get private key" $ \f-> do
+            (ctx, db) <- newDBLayer' (Just f)
+            unsafeRunExceptT $ createWallet db testPk testCp testMetadata
+            (k, h) <- unsafeRunExceptT $ attachPrivateKey db testPk
+            destroyDBLayer ctx
+            testOpeningCleaning f (`readPrivateKey` testPk) (Just (k, h)) Nothing
+
+        it "put and read tx history (Ascending)" $ \f -> do
+            (ctx, db) <- newDBLayer' (Just f)
+            unsafeRunExceptT $ createWallet db testPk testCp testMetadata
+            unsafeRunExceptT $ putTxHistory db testPk (Map.fromList testTxs)
+            destroyDBLayer ctx
+            testOpeningCleaning
+                f
+                (\db' -> readTxHistory db' testPk Ascending wholeRange)
+                testTxs
+                mempty
+
+        it "put and read tx history (Decending)" $ \f -> do
+            (ctx, db) <- newDBLayer' (Just f)
+            unsafeRunExceptT $ createWallet db testPk testCp testMetadata
+            unsafeRunExceptT $ putTxHistory db testPk (Map.fromList testTxs)
+            destroyDBLayer ctx
+            testOpeningCleaning
+                f
+                (\db' -> readTxHistory db' testPk Descending wholeRange)
+                testTxs
+                mempty
+
+        it "put and read checkpoint" $ \f -> do
+            (ctx, db) <- newDBLayer' (Just f)
+            unsafeRunExceptT $ createWallet db testPk testCp testMetadata
+            unsafeRunExceptT $ putCheckpoint db testPk testCp
+            destroyDBLayer ctx
+            testOpeningCleaning f (`readCheckpoint` testPk) (Just testCp) Nothing
+
+    describe "random operation chunks property" $ do
+        it "realize a random batch of operations upon one db open"
+            (property $ prop_randomOpChunks @(SeqState DummyTarget) @DummyTarget)
+
+-- This property checks that executing series of wallet operations in a single
+-- SQLite session has the same effect as executing the same operations over
+-- multiple sessions.
+prop_randomOpChunks
+    :: (Eq s, IsOurs s, NFData s, Show s, PersistState s, PersistTx t)
+    => KeyValPairs (PrimaryKey WalletId) (Wallet s t, WalletMetadata)
+    -> Property
+prop_randomOpChunks (KeyValPairs pairs) =
+    not (null pairs) ==> monadicIO (liftIO prop)
+  where
+    prop = do
+        filepath <- temporaryDBFile
+        (ctxF, dbF) <- newDBLayer' (Just filepath) >>= cleanDB'
+        (ctxM, dbM) <- inMemoryDBLayer >>= cleanDB'
+        forM_ pairs (insertPair dbM)
+        cutRandomly pairs >>= mapM_ (\chunk -> do
+            (ctx, db) <- newDBLayer' (Just filepath)
+            forM_ chunk (insertPair db)
+            destroyDBLayer ctx)
+        dbF `shouldBeConsistentWith` dbM
+        destroyDBLayer ctxF *> destroyDBLayer ctxM
+
+    insertPair
+        :: DBLayer IO s t k
+        -> (PrimaryKey WalletId, (Wallet s t, WalletMetadata))
+        -> IO ()
+    insertPair db (k, (cp, meta)) = do
+        keys <- listWallets db
+        if k `elem` keys then do
+            unsafeRunExceptT $ putCheckpoint db k cp
+            unsafeRunExceptT $ putWalletMeta db k meta
+        else do
+            unsafeRunExceptT $ createWallet db k cp meta
+            Set.fromList <$> listWallets db `shouldReturn` Set.fromList (k:keys)
+
+    shouldBeConsistentWith :: (Eq s) => DBLayer IO s t k -> DBLayer IO s t k -> IO ()
+    shouldBeConsistentWith db1 db2 = do
+        expectedWalIds <- Set.fromList <$> listWallets db1
+        Set.fromList <$> listWallets db2
+            `shouldReturn` expectedWalIds
+        forM_ expectedWalIds $ \walId -> do
+            expectedCps <- readCheckpoint db1 walId
+            readCheckpoint db2 walId
+                `shouldReturn` expectedCps
+        forM_ expectedWalIds $ \walId -> do
+            expectedMetas <- readWalletMeta db1 walId
+            readWalletMeta db2 walId
+                `shouldReturn` expectedMetas
+
+-- | Test that data is preserved between open / close of the same database and
+-- that cleaning up happens as expected.
+testOpeningCleaning
+    :: (Show s, Eq s)
+    => FilePath
+    -> (DBLayer IO (SeqState DummyTarget) DummyTarget SeqKey -> IO s)
+    -> s
+    -> s
+    -> Expectation
+testOpeningCleaning filepath call expectedAfterOpen expectedAfterClean = do
+    (ctx1, db1) <- newDBLayer' (Just filepath)
+    call db1 `shouldReturn` expectedAfterOpen
+    _ <- cleanDB db1
+    call db1 `shouldReturn` expectedAfterClean
+    destroyDBLayer ctx1
+    (ctx2,db2) <- newDBLayer' (Just filepath)
+    call db2 `shouldReturn` expectedAfterClean
+    destroyDBLayer ctx2
+
+
+-- | Run a test action inside withDBLayer, then check assertions.
+withTestDBFile
+    :: (DBLayer IO (SeqState DummyTarget) DummyTarget SeqKey -> IO ())
+    -> (FilePath -> IO a)
+    -> IO a
+withTestDBFile action expectations = do
+    logConfig <- defaultConfigTesting
+    trace <- setupTrace (Right logConfig) "connectionSpec"
+    withSystemTempFile "spec.db" $ \fp _handle -> do
+        removeFile fp
+        withDBLayer logConfig trace (Just fp) action
+        expectations fp
+
+inMemoryDBLayer
+    :: (IsOurs s, NFData s, Show s, PersistState s, PersistTx t)
+    => IO (SqliteContext, DBLayer IO s t SeqKey)
+inMemoryDBLayer = newDBLayer' Nothing
+
+temporaryDBFile :: IO FilePath
+temporaryDBFile = emptySystemTempFile "cardano-wallet-SqliteFileMode"
+
+newDBLayer'
+    :: (IsOurs s, NFData s, Show s, PersistState s, PersistTx t)
+    => Maybe FilePath
+    -> IO (SqliteContext, DBLayer IO s t SeqKey)
+newDBLayer' fp = do
+    logConfig <- CM.empty
+    newDBLayer logConfig nullTracer fp
+
+-- | Clean the database
+cleanDB'
+    :: Monad m
+    => (SqliteContext, DBLayer m s t k)
+    -> m (SqliteContext, DBLayer m s t k)
+cleanDB' (ctx, db) =
+    cleanDB db $> (ctx, db)
+
+-- | Attach an arbitrary private key to a wallet
+attachPrivateKey
+    :: DBLayer IO s t SeqKey
+    -> PrimaryKey WalletId
+    -> ExceptT ErrNoSuchWallet IO (SeqKey 'RootK XPrv, Hash "encryption")
+attachPrivateKey db wid = do
+    let Right pwd = fromText "simplevalidphrase"
+    let k = generateKeyFromSeed (coerce pwd, coerce pwd) pwd
+    h <- liftIO $ encryptPassphrase pwd
+    putPrivateKey db wid (k, h)
+    return (k, h)
+
+cutRandomly :: [a] -> IO [[a]]
+cutRandomly = iter []
+  where
+    iter acc rest
+        | L.length rest <= 1 =
+            pure $ L.reverse (rest:acc)
+        | otherwise = do
+            chunksNum <- randomRIO (1, L.length rest)
+            let chunk = L.take chunksNum rest
+            iter (chunk:acc) (L.drop chunksNum rest)
+
+
+{-------------------------------------------------------------------------------
                                    Test data
 -------------------------------------------------------------------------------}
+
+testCp :: Wallet (SeqState DummyTarget) DummyTarget
+testCp = initWallet block0 genesisParameters initDummyState
+  where
+    initDummyState :: SeqState DummyTarget
+    initDummyState = mkSeqState (xprv, mempty) defaultAddressPoolGap
+      where
+        bytes = entropyToBytes <$> unsafePerformIO $ genEntropy @(EntropySize 15)
+        xprv = generateKeyFromSeed (Passphrase bytes, mempty) mempty
 
 testMetadata :: WalletMetadata
 testMetadata = WalletMetadata
@@ -394,9 +622,10 @@ testPk1 = PrimaryKey testWid1
 
 testTxs :: [(Hash "Tx", (Tx, TxMeta))]
 testTxs =
-    [ (Hash "tx2"
-      , (Tx [TxIn (Hash "tx1") 0] [TxOut (Address "addr") (Coin 1)]
-        , TxMeta InLedger Incoming (SlotId 14 0) (Quantity 1337144))) ]
+    [ (Hash "tx2", (Tx [TxIn (Hash "tx1") 0] [TxOut (Address "addr") (Coin 1)]
+      , TxMeta InLedger Incoming (SlotId 14 0) (Quantity 1337144))
+      )
+    ]
 
 testPassphraseAndHash :: IO (Passphrase "encryption", Hash "encryption")
 testPassphraseAndHash = do
